Remove deprecated overview page layout
