+++ frontend/public/locales/en/translation.json
@@ -12,8 +12,6 @@
   "{{action}} {{resourceKind}}": "{{action}} {{resourceKind}}",
   "{{count}} cluster cannot be edited ": "{{count}} cluster cannot be edited ",
   "{{count}} cluster cannot be edited _plural": "{{count}} clusters cannot be edited ",
-  "{{count}} clusters with issues": "{{count}} cluster with issues",
-  "{{count}} clusters with issues_plural": "{{count}} clusters with issues",
   "{{count}} clusters with unknown status": "{{count}} cluster with unknown status",
   "{{count}} clusters with unknown status_plural": "{{count}} clusters with unknown status",
   "{{count}} clusters with unsatisfied dependencies": "{{count}} cluster with unsatisfied dependencies",
@@ -680,7 +678,6 @@
   "Cluster IP": "Cluster IP",
   "Cluster is offline": "Cluster is offline",
   "Cluster is pending import": "Cluster is pending import",
-  "Cluster issues": "Cluster issues",
   "Cluster Labels": "Cluster Labels",
   "Cluster labels and placement rules": "Cluster labels and placement rules",
   "Cluster landscape": "Cluster landscape",
@@ -698,7 +695,6 @@
   "Cluster Selector": "Cluster Selector",
   "Cluster set bind": "Cluster set bind",
   "Cluster sets": "Cluster sets",
-  "Cluster status": "Cluster status",
   "Cluster update risks are only collected for OpenShift Container Platform clusters.": "Cluster update risks are only collected for OpenShift Container Platform clusters.",
   "Cluster version": "Cluster version",
   "Cluster violation": "Cluster violation",
@@ -781,7 +777,6 @@
   "Compute Optimized": "Compute Optimized",
   "Condition": "Condition",
   "Conditions": "Conditions",
-  "Configuration alert": "Configuration alert",
   "Configuration for disconnected installation": "Configuration for disconnected installation",
   "Configuration objects": "Configuration objects",
   "Configuration Policy": "Configuration Policy",
@@ -1124,7 +1119,6 @@
   "Deployment available": "Deployment available",
   "Deployments unavailable": "Deployments unavailable",
   "deprecated": "Deprecated",
-  "Deprecated": "Deprecated",
   "Description": "Description",
   "Description (120 character limit)": "Description (120 character limit)",
   "description.ansible.job.status": "Ansible Tower Job status",
@@ -1400,13 +1394,11 @@
   "Export unsuccessful": "Export unsuccessful",
   "External": "External",
   "Extra variables": "Extra variables",
-  "Failed": "Failed",
   "Failed to communicate with the server because of a network timeout.": "Failed to communicate with the server because of a network timeout.",
   "Failed to fetch kubeconfig": "Failed to fetch kubeconfig",
   "Failed to get job names from Ansible": "Failed to get job names from Ansible",
   "Failing operators": "Failing operators",
   "False": "False",
-  "Feature deprecation": "Feature deprecation",
   "Feedback": "Feedback",
   "Filter": "Filter",
   "Filters": "Filters",
@@ -1417,7 +1409,6 @@
   "find.label": "Find",
   "Fit to Screen": "Fit to Screen",
   "Fix exceptions": "Fix exceptions",
-  "Fleet view": "Fleet view",
   "Flux": "Flux",
   "Follow documentation for more information.": "Follow documentation for more information.",
   "For more details and logs, click on the nodes to open the properties view.": "For more details and logs, click on the nodes to open the properties view.",
@@ -1658,7 +1649,6 @@
   "Jump to the bottom": "Jump to the bottom",
   "Kind": "Kind",
   "Kubernetes": "Kubernetes",
-  "Kubernetes type": "Kubernetes type",
   "Kubernetes version": "Kubernetes version",
   "Label": "Label",
   "Label expressions": "Label expressions",
@@ -2016,9 +2006,7 @@
   "Overview": "Overview",
   "Overview of available upgrades": "Overview of available upgrades",
   "Overview of cluster add-ons": "Overview of cluster add-ons",
-  "Overview of cluster issues": "Overview of cluster issues",
   "Overview of cluster status": "Overview of cluster status",
-  "Overview of pod count and status": "Overview of pod count and status",
   "Overview of policy set violations": "Overview of policy set violations",
   "Overview of policy violation status": "Overview of policy violation status",
   "Overview of policy violations": "Overview of policy violations",
@@ -2098,7 +2086,6 @@
   "Pod selector": "Pod selector",
   "Pod Selector": "Pod Selector",
   "pods": "pods",
-  "Pods": "Pods",
   "Policies": "Policies",
   "policies.noviolations": "{{count}} policy with no violations",
   "policies.noviolations_plural": "{{count}} policies with no violations",
@@ -2320,7 +2307,6 @@
   "Run an OpenShift cluster where the control plane is decoupled from the data plane, and is treated like a multi-tenant workload on a hosting service cluster. The data plane is on a separate network domain that allows segmentation between management and workload traffic.": "Run an OpenShift cluster where the control plane is decoupled from the data plane, and is treated like a multi-tenant workload on a hosting service cluster. The data plane is on a separate network domain that allows segmentation between management and workload traffic.",
   "Run command": "Run command",
   "Run the following `oc login` command": "Run the following `oc login` command",
-  "Running": "Running",
   "S3 Bucket": "S3 Bucket",
   "save": "Save",
   "Save": "Save",
@@ -2627,7 +2613,6 @@
   "Successful": "Successful",
   "Successfully copied to clipboard!": "Successfully copied to clipboard!",
   "Suggested search templates": "Suggested search templates",
-  "Summary": "Summary",
   "summary.applications": "Applications",
   "summary.cluster.issues": "Identified issues",
   "summary.cluster.issues.description.count": "{{criticalCount}} Critical, {{importantCount}} Important, {{moderateCount}} Moderate, {{lowCount}} Low",
@@ -2832,7 +2817,6 @@
   "The impact of the problem would be {{totalRisk}} if it occurred": "The impact of the problem would be {{totalRisk}} if it occurred",
   "The instance profile for the node pool.": "The instance profile for the node pool.",
   "The instance type for the node pool.": "The instance type for the node pool.",
-  "The layout of this Overview page is deprecated. Enable the Fleet view switch to view the new default overview page.": "The layout of this Overview page is deprecated. Enable the Fleet view switch to view the new default overview page.",
   "The length of application set name must not exceed 53 characters": "The length of application set name must not exceed 53 characters",
   "The location of the Red Hat Enterprise Linux CoreOS image in your local registry.": "The location of the Red Hat Enterprise Linux CoreOS image in your local registry.",
   "The location of the resources on the Git repository.": "The location of the resources on the Git repository.",
@@ -2874,7 +2858,6 @@
   "The resource was not found.": "The resource was not found.",
   "The root volume size for the node pool.": "The root volume size for the node pool.",
   "The root volume type for the node pool.": "The root volume type for the node pool.",
-  "The search service is unavailable or degraded. Some data might be missing from this view.": "The search service is unavailable or degraded. Some data might be missing from this view.",
   "The specific name for the target Helm chart.": "The specific name for the target Helm chart.",
   "The statues in this resource group have not been found and are unknown.": "The statues in this resource group have not been found and are unknown.",
   "The subject name should match the name of a policy or policy set in this namespace.": "The subject name should match the name of a policy or policy set in this namespace.",
