+++ frontend/src/routes/Home/Overview/components/SavedSearchesCard.tsx
@@ -17,13 +17,13 @@ import {
 import { CogIcon, ExclamationCircleIcon } from '@patternfly/react-icons'
 import { Fragment, useEffect, useState } from 'react'
 import { Link } from 'react-router-dom-v5-compat'
-import { useTranslation } from '../../../lib/acm-i18next'
-import { NavigationPath } from '../../../NavigationPath'
-import { IResource, listResources, SavedSearch } from '../../../resources'
-import { useSharedAtoms } from '../../../shared-recoil'
-import { convertStringToQuery } from '../../Search/search-helper'
-import { searchClient } from '../../Search/search-sdk/search-client'
-import { useSearchResultCountQuery } from '../../Search/search-sdk/search-sdk'
+import { useTranslation } from '../../../../lib/acm-i18next'
+import { NavigationPath } from '../../../../NavigationPath'
+import { IResource, listResources, SavedSearch } from '../../../../resources'
+import { useSharedAtoms } from '../../../../shared-recoil'
+import { convertStringToQuery } from '../../../Search/search-helper'
+import { searchClient } from '../../../Search/search-sdk/search-client'
+import { useSearchResultCountQuery } from '../../../Search/search-sdk/search-sdk'
 
 const CardHeader = (props: { isSearchDisabled: boolean }) => {
   const { isSearchDisabled } = props
