+++ frontend/src/routes/Home/Overview/components/SummaryCard.tsx
@@ -16,10 +16,10 @@ import {
   Title,
 } from '@patternfly/react-core'
 import { ExclamationCircleIcon } from '@patternfly/react-icons'
-import { Link } from 'react-router-dom-v5-compat'
 import { ReactNode } from 'react'
-import { useTranslation } from '../../../lib/acm-i18next'
-import { AcmButton } from '../../../ui-components'
+import { Link } from 'react-router-dom-v5-compat'
+import { useTranslation } from '../../../../lib/acm-i18next'
+import { AcmButton } from '../../../../ui-components'
 
 function getLink(type: 'link' | 'button', path: string, count: number) {
   if (type === 'link') {
